// apps/sales-copilot/src/generated/models/ActivitymimeattachmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Activitymimeattachmentscomponentstate = {
  0: 'Published',
  1: 'Unpublished',
  2: 'Deleted',
  3: 'DeletedUnpublished'
} as const;
export type Activitymimeattachmentscomponentstate = keyof typeof Activitymimeattachmentscomponentstate;
export const Activitymimeattachmentsobjecttypecode = {
  2010: 'EmailTemplate',
  4200: 'EmailActivity'
} as const;
export type Activitymimeattachmentsobjecttypecode = keyof typeof Activitymimeattachmentsobjecttypecode;

export interface ActivitymimeattachmentsBase {
  "ActivityId@odata.bind"?: string;
  activitymimeattachmentid: string;
  activitymimeattachmentidunique: string;
  attachmentcontentid?: string;
  "AttachmentId@odata.bind": string;
  attachmentnumber: number;
  body?: string;
  filename?: string;
  mimetype?: string;
  "ObjectId@odata.bind": string;
  objecttypecode: Activitymimeattachmentsobjecttypecode;
  subject?: string;
}

export interface Activitymimeattachments extends ActivitymimeattachmentsBase {
  activitysubject?: string;
  anonymouslink?: string;
  componentstate: Activitymimeattachmentscomponentstate;
  filesize?: number;
  isfollowed?: boolean;
  isfollowedname?: string;
  ismanaged: boolean;
  ismanagedname?: string;
  overwritetime: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  solutionid: string;
  supportingsolutionid?: string;
  versionnumber?: number;
  activityid?: object;
  _activityid_value?: string;
  attachmentid?: object;
  _attachmentid_value?: string;
  objectid?: object;
  _objectid_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}